Added catalog testDB Added test.py Deleted test from main.py
